[ADD] asd_base: init
