+++ asd_base/__manifest__.py
@@ -2,14 +2,14 @@
 # Copyright 2024 PT Arkana Solusi Digital
 # License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl.html).
 {
-    "name": "Pennyu - Company",
-    "summary": """Addition of company master data configuration""",
+    "name": "Pennyu - Base",
+    "summary": """Additional general features related to the Base module""",
     "author": "PT Arkana Solusi Digital",
     "website": "https://arkana.co.id",
     "category": "Uncategorized",
     "version": "0.1",
     "depends": ["base"],
     "data": [
-        "views/res_company_view.xml",
+        "views/res_partner_view.xml",
     ],
 }
