--- asd_company/views/res_company_view.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-  <data>
-    <record model="ir.ui.view" id="view_company_form">
-      <field name="name">res.company.form</field>
-      <field name="model">res.company</field>
-      <field name="inherit_id" ref="base.view_company_form"/>
-      <field name="arch" type="xml">
-        <xpath expr="//field[@name='report_header']" position="before">
-          <field name="show_id_sfa" invisible="1"/>
-          <field name="id_sfa" attrs="{'invisible': [('show_id_sfa','!=',True)]}"/>
-        </xpath>
-      </field>
-    </record>
-  </data>
-</odoo>
